Whats New!
Micro:Radio V3 Features UpdateCheck which communicates with other Micro:BITS to see if you are on the correct version. It also features quicker startup times, and various bug fixes and also not being able to disconect when SOS mode is on and also made turning off SOS mode less annoying.